# api/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "This is the FAIMS Conductor server API documentation.",
    "version": "1.0.0",
    "title": "FAIMS3 API",
    "termsOfService": "http://faims.edu.au/terms/",
    "contact": {"email": "help@faims.edu.au"},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [{"url": "/api"}],
  "components": {
    "securitySchemes": {
      "Auth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}
    },
    "responses": {
      "UnauthorizedError": {"description": "Access token is missing or invalid"}
    }
  },
  "paths": {
...
    "/invites/notebook/{projectId}": {
      "get": {
        "summary": "Get all project invites",
        "tags": ["Invites"],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "security": [{"Auth": []}],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/definitions/GetProjectInvitesResponse"}
              }
            }
          },
          "401": {"$ref": "#/components/responses/UnauthorizedError"}
        }
      },
      "post": {
        "summary": "Create a project invite",
        "description": "Creates a new invite for a project/notebook. Requires appropriate permissions.",
        "tags": ["Invites"],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {"type": "string"},
            "description": "Project/notebook identifier"
          }
        ],
        "security": [{"Auth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/definitions/PostCreateInviteInput"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/PostCreateProjectInviteResponse"
                }
              }
            }
          },
          "401": {"$ref": "#/components/responses/UnauthorizedError"}
        }
      }
    },
    "/invites/team/{teamId}": {
      "get": {
        "summary": "Get all team invites",
        "tags": ["Invites"],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "security": [{"Auth": []}],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/definitions/GetTeamInvitesResponse"}
              }
            }
          },
          "401": {"$ref": "#/components/responses/UnauthorizedError"}
        }
      },
      "post": {
        "summary": "Create a team invite",
        "description": "Creates a new invite for a team. Requires appropriate permissions.",
        "tags": ["Invites"],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {"type": "string"},
            "description": "Team identifier"
          }
        ],
        "security": [{"Auth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/definitions/PostCreateInviteInput"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/definitions/PostCreateTeamInviteResponse"}
              }
            }
          },
          "401": {"$ref": "#/components/responses/UnauthorizedError"}
        }
      }
    },
...
  },
  "definitions": {
...
    "GetProjectInvitesResponse": {
      "type": "object",
      "properties": {
        "invites": {
          "type": "array",
          "items": {"$ref": "#/definitions/ProjectInviteDocument"}
        }
      }
    },
    "GetTeamInvitesResponse": {
      "type": "object",
      "properties": {
        "invites": {
          "type": "array",
          "items": {"$ref": "#/definitions/TeamInviteDocument"}
        }
      }
    },
    "ProjectInviteDocument": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "description": "Unique identifier for the invite"
        },
        "_rev": {
          "type": "string",
          "description": "CouchDB revision identifier"
        },
        "projectId": {
          "type": "string",
          "description": "Project/notebook identifier"
        },
        "email": {
          "type": "string",
          "format": "email",
          "description": "Email address of the invitee"
        },
        "role": {
          "type": "string",
          "description": "Role to be assigned to the user"
        },
        "invitedBy": {
          "type": "string",
          "description": "User ID who created the invite"
        },
        "createdAt": {
          "type": "number",
          "description": "Unix timestamp when invite was created"
        },
        "expiresAt": {
          "type": "number",
          "description": "Unix timestamp when invite expires"
        },
        "status": {
          "type": "string",
          "enum": ["pending", "accepted", "expired", "revoked"],
          "description": "Status of the invite"
        }
      },
      "required": [
        "_id",
        "projectId",
        "email",
        "role",
        "invitedBy",
        "createdAt",
        "status"
      ]
    },
    "TeamInviteDocument": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "description": "Unique identifier for the invite"
        },
        "_rev": {
          "type": "string",
          "description": "CouchDB revision identifier"
        },
        "teamId": {"type": "string", "description": "Team identifier"},
        "email": {
          "type": "string",
          "format": "email",
          "description": "Email address of the invitee"
        },
        "role": {
          "type": "string",
          "description": "Role to be assigned to the user in the team"
        },
        "invitedBy": {
          "type": "string",
          "description": "User ID who created the invite"
        },
        "createdAt": {
          "type": "number",
          "description": "Unix timestamp when invite was created"
        },
        "expiresAt": {
          "type": "number",
          "description": "Unix timestamp when invite expires"
        },
        "status": {
          "type": "string",
          "enum": ["pending", "accepted", "expired", "revoked"],
          "description": "Status of the invite"
        }
      },
      "required": [
        "_id",
        "teamId",
        "email",
        "role",
        "invitedBy",
        "createdAt",
        "status"
      ]
    },
    "PostCreateInviteInput": {
      "type": "object",
      "properties": {
        "role": {
          "type": "string",
          "description": "Role to assign to the invited user"
        },
        "name": {
          "type": "string",
          "description": "Name/email of the person being invited"
        },
        "expiry": {
          "type": "number",
          "description": "Unix timestamp when invite expires (optional)"
        },
        "uses": {
          "type": "number",
          "description": "Number of times the invite can be used (optional)"
        }
      },
      "required": ["role", "name"]
    },
    "PostCreateProjectInviteResponse": {
      "$ref": "#/definitions/ProjectInviteDocument"
    },
    "PostCreateTeamInviteResponse": {
      "$ref": "#/definitions/TeamInviteDocument"
    },
...
  }
}